pv-emulator/requirements.txt
dash>=2.16.0
dash-bootstrap-components>=1.5.0
plotly>=5.20.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
requests>=2.31.0
pyserial>=3.5
pvlib>=0.10.0
pyyaml>=6.0
pywebview>=4.4.0   # opcional — modo escritorio sin navegador
